// src/shared/api/generated/model/reportsGetCenterReportPeriod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */

export type ReportsGetCenterReportPeriod = typeof ReportsGetCenterReportPeriod[keyof typeof ReportsGetCenterReportPeriod];


export const ReportsGetCenterReportPeriod = {
  week: 'week',
  month: 'month',
  quarter: 'quarter',
} as const;