// src/shared/api/generated/model/centerReportResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { CenterReportResponseDtoIncomeByMonthItem } from './centerReportResponseDtoIncomeByMonthItem.ts';
import type { CenterReportResponseDtoPeriod } from './centerReportResponseDtoPeriod.ts';
import type { CenterReportResponseDtoRetentionByJoinMonthItem } from './centerReportResponseDtoRetentionByJoinMonthItem.ts';
import type { CenterReportResponseDtoServicesItem } from './centerReportResponseDtoServicesItem.ts';
import type { CenterReportResponseDtoStaffItem } from './centerReportResponseDtoStaffItem.ts';

export interface CenterReportResponseDto {
  period: CenterReportResponseDtoPeriod;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$ */
  fromDate: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$ */
  toDate: string;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  estimatedIncomeCents: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  previousEstimatedIncomeCents: number;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  averageOccupancyPercent: number | null;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  retentionThreeMonthsPercent: number | null;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  activeClientCount: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  inactiveClientCount: number;
  incomeByMonth: CenterReportResponseDtoIncomeByMonthItem[];
  services: CenterReportResponseDtoServicesItem[];
  staff: CenterReportResponseDtoStaffItem[];
  retentionByJoinMonth: CenterReportResponseDtoRetentionByJoinMonthItem[];
}